requirements.txt
opencv_python_headless==4.8.1.78
streamlit==1.29.0
mediapipe==0.10.9
numpy==1.26.2
Pillow==10.1.0
